display cost summary display total cost/days in project summary refresh project on save

+++ Chiffrage.Projects.Module/ViewModel/ProjectViewModel.cs
@@ -33,7 +33,7 @@ namespace Chiffrage.Projects.Module.ViewModel
 
         public virtual double TestNightRate { get; set; }
 
-        public virtual int TotalDays { get; set; }
+        public virtual double TotalDays { get; set; }
 
         public virtual double TotalPrice { get; set; }
